pyproject.toml
[project]
name = "models4fusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dm-haiku>=0.0.16",
    "flax>=0.12.2",
    "jax[tpu]>=0.8.2",
    "mlp-gpt-jax>=0.0.20",
]
